refactor: separa rotas do agent por arquivo e organiza estrutura do projeto
