+++ nutrixpert/agent.py
@@ -1,7 +1,6 @@
 import os
 from dotenv import load_dotenv
 from google.adk.agents import Agent, LlmAgent
-from google.adk.models.lite_llm import LiteLlm
 from nutrixpert.core.tools import (
     calc_macros_tool,
     retrieve_context_tool,
@@ -26,13 +25,10 @@ ADK_MODEL = os.getenv("ADK_MODEL")
 
 def build_nutricional_agent() -> Agent:
     """Agente especialista em composição e substituições alimentares"""
-    ollama_medgemma = LiteLlm(
-        model=ADK_MODEL
-    )
 
     return LlmAgent(
         name="Agente_Nutricional",
-        model=ollama_medgemma,
+        model=ADK_MODEL,
         instruction=AGENT_NUTRICAO_INSTR,
         output_key=AGENT_OUTPUT_KEY,
         tools=[
@@ -48,7 +44,7 @@ def build_metabolico_agent() -> Agent:
     """Agente responsável por cálculos de metabolismo e TMB"""
     return LlmAgent(
         name="Agente_Metabolico",
-        model="gemini-2.0-flash", # modelo especializado médico
+        model=ADK_MODEL,
         instruction=AGENT_METABOLICO_INSTR,
         output_key=AGENT_OUTPUT_KEY,
         tools=[
@@ -63,7 +59,7 @@ def build_planejamento_agent() -> Agent:
     """Agente para planejamento de cardápios"""
     return LlmAgent(
         name="Agente_Planejamento",
-        model="gemini-2.0-flash",
+        model=ADK_MODEL,
         instruction=AGENT_PLANEJAMENTO_INSTR,
         output_key=AGENT_OUTPUT_KEY,
         tools=[
@@ -77,7 +73,7 @@ def build_educativo_agent() -> Agent:
     """Agente que responde dúvidas teóricas sobre nutrição"""
     return LlmAgent(
         name="Agente_Educativo",
-        model="gemini-2.0-flash",
+        model=ADK_MODEL,
         instruction=AGENT_EDUCATIVO_INSTR,
         output_key=AGENT_OUTPUT_KEY,
         tools=[
@@ -96,7 +92,7 @@ def build_root_agent() -> Agent:
 
     root_agent = LlmAgent(
         name=ADK_APP_NAME,
-        model="gemini-2.0-flash",
+        model=ADK_MODEL,
         instruction=ROOT_AGENT_INSTR,
         description="Gerencia o fluxo entre subagentes de nutrição",
         sub_agents=[
