+++ nutrixpert/core/prompt/AGENT_METABOLICO_INSTR.py
@@ -19,8 +19,7 @@ REGRAS:
 - Sempre cite as fórmulas e explique brevemente como o cálculo é feito.
 - Retorne resultados em kcal e g.
 
-Se receber uma pergunta fora da sua especialidade, **responda você mesmo** da melhor forma possível,
-ou encaminhe internamente para outro agente **sem mencionar transferência ou outros agentes**.
+Se receber uma pergunta fora da sua especialidade, encaminhe internamente para outro agente **sem mencionar transferência ou outros agentes**.
 Nunca diga frases como:
 - "Essa pergunta seria melhor respondida por outro agente"
 - "Prefere que eu o transfira"
