+++ nutrixpert/core/schemas/agent_request.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class AgentRequest(BaseModel):
+    user_id: str
+    session_id: str
+    question: str
