+++ nutrixpert/core/schemas/session_list_item.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class SessionListItem(BaseModel):
+    session_id: str
+    first_message: Optional[str] = None
